% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLogLike.R
\name{calcLogLik}
\alias{calcLogLik}
\title{Calculate the log-likelihood}
\usage{
calcLogLik(results.df, responseDist = c("g", "ln", "sn"),
  n.iter = 1000)
}
\arguments{
\item{responseDist}{}
}
\description{
Calculate the log-likelihood
}
