% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLL.R
\name{compareLL}
\alias{compareLL}
\title{Compare log-likelihoods}
\usage{
compareLL(model1LL.df, model2LL.df, model1lab = NULL, model2lab = NULL)
}
\arguments{
\item{model1LL.df}{data frame containing log likelihood from model 1 - comes from calcLogLik}

\item{model2LL.df}{data frame containing log likelihood from model 2 - comes from calcLogLik}

\item{model1lab}{optional label for model 1}

\item{model2lab}{optional label for model 2}
}
\description{
Compare log-likelihoods
}
