% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsPlot.R
\name{effectsPlot}
\alias{effectsPlot}
\title{Plot of the error bars for the effects}
\usage{
effectsPlot(results.df, effect = c("locus", "profile", "dye",
  "locus.dye"), xlab = NULL, ylab = NULL)
}
\arguments{
\item{effect}{}
}
\description{
Plot of the error bars for the effects
}
