% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadResults.R
\name{loadResults}
\alias{loadResults}
\title{Load results file}
\usage{
loadResults(model = c(paste0("g-", 0:4), paste0("ln-", 0:9),
  paste0("sn-", 0:4)), resultsRoot = "../systematic/", summary = FALSE)
}
\arguments{
\item{summary}{}
}
\description{
Provides a simple way to load results files and turns
them into a data.frame that can be manipulated
}
