% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBUGSinputs.R
\name{makeBUGSinputs}
\alias{makeBUGSinputs}
\title{Make BUGS inputs}
\usage{
makeBUGSinputs(form = formula("y ~ 1"), data.df,
  responseDist = c("gamma", "normal", "sn"), genInits = FALSE)
}
\arguments{
\item{form}{a formula}

\item{data.df}{a data.frame}

\item{responseDist}{choice of the peak heights being gamma or (log)-normal}
}
\value{
a list with some stuff in it
}
\description{
A simple mechanism for selecting the set of models we want to explore
}
\examples{
data.df = readData()
makeBUGSinputs(data = data.df)
}
