% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsExpectPlot.R
\name{obsExpectPlot}
\alias{obsExpectPlot}
\title{Plot the Expected versus Observed or Pred-Res}
\usage{
obsExpectPlot(results.df, responseDist = c("g", "ln", "sn"),
  predRes = FALSE, standardize = FALSE)
}
\arguments{
\item{results.df}{A data frame from the loadResults function}

\item{responseDist}{one of \code{"g"},\code{"ln"}, or \code{"sn"}}

\item{predRes}{if \code{TRUE} then a pred-res (residuals vs. fitted) plot is displayed}
}
\value{
a list with the ggplot object and the data used to create the plot
}
\description{
Plot the Expected versus Observed or Pred-Res
}
