% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRawData.R
\name{readRawData}
\alias{readRawData}
\title{readRawData}
\usage{
readRawData(LocusNames = c("K1", "M1", "R1", "K2", "Y1", "M2", "R2",
  "Y2", "Y3", "R3", "R4", "Y4", "Y5", "R5", "Y6", "H1", "B1", "S1", "V1",
  "R6", "Y7", "Y8", "R7", "S2", "R8", "Y9", "S3", "R9", "Y10", "R10",
  "Y11"))
}
\arguments{
\item{LocusNames}{}
}
\value{
a list
}
\description{
Takes the original data and turns it into a data structure than can be used with rjags.
You shouldn't need to call this directly.
}
\examples{
BUGSdata = readRawData()
}
