% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSim.R
\name{runSim}
\alias{runSim}
\title{Run a simulation/fit the model}
\usage{
runSim(form, data, simPath, simRoot, responseDist = c("gamma", "normal",
  "sn"), genInits = FALSE, nUpdate = 10000, nChains = 1,
  nIter = 1000)
}
\description{
Run a simulation/fit the model
}
