% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLL.R
\name{compareLL}
\alias{compareLL}
\title{Compare log-likelihoods}
\usage{
compareLL(..., labels = NULL)
}
\arguments{
\item{\dots}{two or more data frames containing log likelihoods which come from calcLogLik}

\item{lables}{an optional vector of label for each model}
}
\description{
Compare log-likelihoods
}
